+++ las-analyzer/src/main/java/org/easycloud/las/analyzer/recommender/item/PartialMultiplyMapper.java
@@ -32,13 +32,13 @@ public class PartialMultiplyMapper extends MapReduceBase
         }
         RandomAccessVector<String> cooccurrenceColumn = value.getVector();
         List<String> userIDs = value.getUserIDs();
-        List<Double> prefValues = value.getValues();
+        List<Integer> prefValues = value.getValues();
         VectorWritable outputValue = new VectorWritable();
 
         for (int i = 0; i < userIDs.size(); i++) {
             String userID = userIDs.get(i);
-            double prefValue = prefValues.get(i);
-            if (prefValue > 0.0) {
+            int prefValue = prefValues.get(i);
+            if (prefValue > 0) {
                 RandomAccessVector<String> partialProduct = cooccurrenceColumn.times(prefValue);
                 outputValue.set(partialProduct);
                 collector.collect(new Text(userID), outputValue);
